# src/main/java/com/creditsvc/calculator/swagger/api/CreditserviceApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.25).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.creditsvc.calculator.swagger.api;

import java.util.List;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.creditsvc.calculator.swagger.model.CalculatorRequest;
import com.creditsvc.calculator.swagger.model.CalculatorResponse;
import com.creditsvc.calculator.swagger.model.ErrorResponse;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2021-05-02T18:59:50.045Z[GMT]")
@Validated
public interface CreditserviceApi {

    @Operation(summary = "Retrieve a credit assessment score for a given business demographic.", description = "" 
/*    		,security = {
        @SecurityRequirement(name = "creditsvc_auth", scopes = {
            "write:pets",
"read:pets"        })    } 
*/
    , tags={ "Credit Assessment" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "successful operation", content = @Content(array = @ArraySchema(schema = @Schema(implementation = CalculatorResponse.class)))),
        
        @ApiResponse(responseCode = "400", description = "Invalid input", content = @Content(array = @ArraySchema(schema = @Schema(implementation = ErrorResponse.class)))),
        
        @ApiResponse(responseCode = "401", description = "Not authorized", content = @Content(array = @ArraySchema(schema = @Schema(implementation = ErrorResponse.class)))),
        
        @ApiResponse(responseCode = "403", description = "Not authenticated", content = @Content(array = @ArraySchema(schema = @Schema(implementation = ErrorResponse.class)))),
        
        @ApiResponse(responseCode = "404", description = "Api not found", content = @Content(array = @ArraySchema(schema = @Schema(implementation = ErrorResponse.class)))) })
    @RequestMapping(value = "/calculator",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<CalculatorResponse> retrieveScore(@Parameter(in = ParameterIn.DEFAULT, description = "Retrieve a credit assessment score", required=true, schema=@Schema()) @Valid @RequestBody CalculatorRequest body);

}